+++ theme-generator/themes/generate-dark-theme.ts
@@ -1422,6 +1422,10 @@ export const generateDarkTheme = (
         textColor,
       },
     },
+    viewport: {
+      backgroundColor: overlayBackgroundColor,
+      textColor,
+    },
     welcomeScreen: {
       backgroundColor,
       borderColor: guidelinesColor,